expose SceneData and from the types file
