+++ react-components/src/hooks/index.ts
@@ -17,10 +17,13 @@ export { use3dScenes } from './scenes/use3dScenes';
 export { useSceneConfig } from './scenes/useSceneConfig';
 export { useGhostMode } from './useGhostMode';
 
+export type { SceneData } from './scenes/types';
+
 export type { CameraNavigationActions } from './useCameraNavigation';
 export type { ClickedNodeData, FdmNodeDataResult } from './useClickedNode';
 export type {
   PointCloudAnnotationMappedAssetData,
   Image360AnnotationMappedAssetData,
-  ThreeDModelFdmMappings
+  ThreeDModelFdmMappings,
 } from './types';
+
